order fixed

+++ models/order.go
@@ -26,7 +26,7 @@ type OrderList struct {
 	Address   *string   `json:"address"`
 	Status    int16     `json:"status"`
 	Square    float64   `json:"square"`
-	Price     int       `json:"price"`
+	Price     float64   `json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
@@ -66,7 +66,7 @@ type Order struct {
 	Longitude             *float64  `json:"longitude"`
 	Address               *string   `json:"address"`
 	Square                float64   `json:"square"`
-	Price                 int       `json:"price"`
+	Price                 float64   `json:"price"`
 }
 
 type OrderSendLocationRequest struct {
